Updated AutoComplete Shortcut to CTRL+SPACE Update to IDE Compile with extra delays to account for spaces Monster Debugger Snippet added
Added new Variable TM_FLEX_CONFIG to allow projects to reference the flex
config file that will be used in the compile. This allows projects
to locate the compile config anywhere, and switch it easily.

Added 3 new Projects
AS3 Standalone is a simple flex sdk based project, running this project
will open the swf in the flash player

AS3 Embeded is the same as AS3 Standalone, but uses the deploy folder
and SWFObject.Running opens the browser for the embeded SWF

Flash IDE Project contains a Flash File in the fla folder with proper
paths setup. Use the Build in Flash IDE to compile this project.

Renamed original project to "Actionscript 3 Project"